+++ AdvancedSharpAdbClient/Exceptions/ExceptionExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
-using System.Runtime.CompilerServices;
-using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.CompilerServices;
 
 namespace AdvancedSharpAdbClient.Exceptions
 {
