+++ AdvancedSharpAdbClient/Logs/ShellStream.cs
@@ -5,7 +5,6 @@
 using AdvancedSharpAdbClient.Exceptions;
 using System;
 using System.IO;
-using System.Runtime.InteropServices;
 using System.Threading;
 
 namespace AdvancedSharpAdbClient.Logs
