+++ AdvancedSharpAdbClient/Receivers/IShellOutputReceiver.cs
@@ -2,8 +2,6 @@
 // Copyright (c) The Android Open Source Project, Ryan Conrad, Quamotion, yungd1plomat, wherewhere. All rights reserved.
 // </copyright>
 
-using System.Threading;
-
 namespace AdvancedSharpAdbClient
 {
     /// <summary>
